fix(blue-section): add folder to move files and imports
