+++ src/pages/ManageValidator/LeftManage/ValidatorBlueSection/ValidatorBlueSection.stories.tsx
