+++ src/pages/ManageValidator/LeftManage/ValidatorBlueSection/ValidatorInfo.tsx
@@ -1,8 +1,10 @@
 import { XStack, YStack } from 'tamagui'
-import ValidatorNameAddress from '../../../components/General/ValidatorNameAddress'
 import { Avatar, Text } from '@status-im/components'
 import { ChevronLeftIcon, ChevronRightIcon } from '@status-im/icons'
-import Icon from '../../../components/General/Icon'
+
+import ValidatorNameAddress from '../../../../components/General/ValidatorNameAddress'
+import Icon from '../../../../components/General/Icon'
+
 const ValidatorInfo = () => {
   return (
     <XStack space={'$2'} padding="30px" justifyContent="space-between">
